% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margins.R
\name{berncpdf}
\alias{berncpdf}
\title{Bernoulli with p = 1/(1+exp(-th)) cdf/pdf and derivatives}
\usage{
berncpdf(z, th)
}
\arguments{
\item{z}{vector of responses}

\item{th}{linear combination of covariates (can be negative)}
}
\value{
\item{out}{Matrix of conditional cdf and pdf with derivative with respect to parameters}
}
\description{
This function computes the cdf, pdf, and associated derivatives
}
\examples{
out = berncpdf(0,2.5)

}
\author{
Pavel Krupskii and Bruno N. Remillard, January 20, 2022
}
