% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coplik.R
\name{coplik}
\alias{coplik}
\title{Copula cdf/pdf and ders}
\usage{
coplik(u, v, family, rot = 0, cpar, dfC = NULL, du = FALSE)
}
\arguments{
\item{u}{vector of values in (0,1)}

\item{v}{conditioning variable in (0,1)}

\item{family}{copula family: "gaussian", "t", "clayton",  "frank", "fgm", "gumbel", "joe", "plackett".}

\item{rot}{rotation: 0 (default), 90, 180 (survival), or 270}

\item{cpar}{copula parameter}

\item{dfC}{degrees of freedom for the Student copula (default is NULL)}

\item{du}{logical value (default = FALSE) for the derivative of the copula density with respect to u}
}
\value{
\item{out}{Matrix of conditional cdf, pdf, and derivatives with respect to parameters}
}
\description{
Derivatives C(u|v), C'_dl(u|v), c(u,v), c'_dl(u,v), c'_u(u,v) for the linking copula
}
\examples{
out = coplik(0.3,0.5,"clayton",cpar=2,du=TRUE)
}
\author{
Pavel Krupskii and Bruno N. Remillard, January 20, 2022
}
