% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcop.R
\name{dbvn}
\alias{dbvn}
\title{Normal density}
\usage{
dbvn(x1, x2, rh)
}
\arguments{
\item{x1}{vector of values}

\item{x2}{vector of values}

\item{rh}{correlation parameter, -1< rh <1}
}
\value{
\item{out}{Vector of densities}
}
\description{
Density at (x1,x2)
}
\examples{
out = dbvn(0.3,0.5,-0.6)


}
\author{
Pavel Krupskii
}
