% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcop.R
\name{dbvncop}
\alias{dbvncop}
\title{Normal copula density}
\usage{
dbvncop(u, v, cpar)
}
\arguments{
\item{u}{vector of values in (0,1)}

\item{v}{vector of values in (0,1)}

\item{cpar}{copula parameter, -1< cpar<1}
}
\value{
\item{out}{Vector of densities}
}
\description{
Density at (u,v)
}
\examples{
out = dbvncop(0.3,0.5,-0.5)
}
\author{
Pavel Krupskii
}
