% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcop.R
\name{dbvtcop}
\alias{dbvtcop}
\title{Student copula density}
\usage{
dbvtcop(u, v, cpar, dfC)
}
\arguments{
\item{u}{vector of values in (0,1)}

\item{v}{vector of values in (0,1)}

\item{cpar}{copula parameter, -1< cpar<1}

\item{dfC}{degrees of freedom}
}
\value{
\item{out}{Vector of densities}
}
\description{
Density at (u,v)
}
\examples{
out = dbvtcop(0.3,0.5,-0.7,25)
}
\author{
Pavel Krupskii
}
