% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcop.R
\name{dcop}
\alias{dcop}
\title{Copula pdf}
\usage{
dcop(u, v, family, rot = 0, cpar, dfC = NULL)
}
\arguments{
\item{u}{vector of values in (0,1)}

\item{v}{conditioning variable in (0,1)}

\item{family}{copula family: "gaussian" ("normal), "t", "clayton",  "frank", "fgm", "galambos", "gumbel", "joe", "huesler-reiss", "plackett".}

\item{rot}{rotation: 0 (default), 90, 180 (survival), or 270}

\item{cpar}{copula parameter}

\item{dfC}{degrees of freedom for the Student copula (default is NULL)}
}
\value{
\item{out}{Copula density}

\item{out}{Vector of pdf values}
}
\description{
Evaluates the copula density at given points (u,v)#'
}
\examples{
out = dcop(0.3,0.7,"clayton",270,2)
}
\author{
Pavel Krupskii and Bruno Remillard, May 1, 2023
}
