% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcop.R
\name{dfrk}
\alias{dfrk}
\title{B3 bivariate Frank copula density}
\usage{
dfrk(u, v, cpar)
}
\arguments{
\item{u}{vector of values in (0,1)}

\item{v}{vector of values in (0,1)}

\item{cpar}{copula parameter, cpar>0 or cpar<0}
}
\value{
\item{out}{Vector of densities}
}
\description{
Density at (u,v)
}
\examples{
out = dfrk(0.3,0.5,2)
}
\author{
Pavel Krupskii
}
