% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcop.R
\name{dgum}
\alias{dgum}
\title{B6 Gumbel copula density, cpar>1}
\usage{
dgum(u, v, cpar)
}
\arguments{
\item{u}{vector of values in (0,1)}

\item{v}{vector of values in (0,1)}

\item{cpar}{copula parameter > 0}
}
\value{
\item{out}{Vector of densities}
}
\description{
Density at (u,v)
}
\examples{
out = dgum(0.3,0.5,2)
}
\author{
Pavel Krupskii
}
