% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margins.R
\name{expcpdf}
\alias{expcpdf}
\title{Exponential cdf/pdf and ders}
\usage{
expcpdf(z, th)
}
\arguments{
\item{z}{vector of responses}

\item{th}{th is rate > 0}
}
\value{
\item{out}{Matrix of conditional cdf, derivative with respect to parameter, pdf, }
}
\description{
This function computes the cdf, pdf, and associated derivatives
}
\examples{
out = expcpdf(2,3)
}
\author{
Pavel Krupskii and Bruno N. Remillard, January 20, 2022
}
