% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margins.R
\name{geomcpdf}
\alias{geomcpdf}
\title{Geometric with p = 1/(1+exp(-th)) cdf/pdf and ders}
\usage{
geomcpdf(z, th)
}
\arguments{
\item{z}{vector of responses}

\item{th}{linear combination of covariates (can be negative)}
}
\value{
\item{out}{Matrix of conditional cdf, derivative with respect to parameter, pdf, }
}
\description{
This function computes the cdf, pdf, and associated derivatives
}
\examples{
out = geomcpdf(0,-3)
}
\author{
Pavel Krupskii and Bruno N. Remillard, January 20, 2022
}
