% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invfunc.R
\name{invfunc}
\alias{invfunc}
\title{Inverse function}
\usage{
invfunc(q, func, th, lb = 1e-12, ub = 1 - 1e-12, tol = 1e-08, nbreak = 40)
}
\arguments{
\item{q}{Function value (can be a vector if func() supports a vector argument)}

\item{func}{Function of one argument to be inverted}

\item{th}{Function parameters}

\item{lb}{Lower bound for the possible values}

\item{ub}{Upper bound for the possible values}

\item{tol}{Tolerance for the inversion}

\item{nbreak}{Maximum number of iterations (default is 40)}
}
\value{
\item{out}{Inverse values}
}
\description{
This function is used to get the inverse of a monotonic function on (0,1), depending on parameters, and using the bisection method
}
\author{
Pavel Krupskii
}
