% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkCop.R
\name{linkCop}
\alias{linkCop}
\title{Link to copula parameter}
\usage{
linkCop(th, family = "clayton")
}
\arguments{
\item{th}{vector of linear combinations}

\item{family}{copula family: "gaussian" , "t" , "clayton" ,   "claytonR" , "frank" , "gumbel", "gumbelR".}
}
\value{
\item{cpar}{Associated copula parameters}

\item{hder}{Derivative of link function}
}
\description{
Computes the copula parameters given a linear combination of covariates.
}
\examples{
out = linkCop(-1,"gaussian")
}
\references{
Krupskii, Nasri & Remillard (2023). On factor copula-based mixed regression models
}
\author{
Pavel Krupskii and Bruno N. Remillard, January 20, 2023
}
