% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margins.R
\name{margins}
\alias{margins}
\title{Margins  cdf/pdf and their derivatives}
\usage{
margins(z, th, model, x = NULL, dfM = NULL)
}
\arguments{
\item{z}{vector of responses}

\item{th}{linear combination of covariates (can be negative)}

\item{model}{model for margin: "binomial" (bernoulli), "poisson", "nbinom" (mean is the parameter),"nbinom1" (p is the parameter), "geometric", "multinomial", "exponential", "weibull", "normal","t", "laplace"}

\item{x}{covariates for the multinomial margin (default is NULL)}

\item{dfM}{degrees of freedom for the Student margin (default is NULL)}
}
\value{
\item{out}{Matrix of conditional cdf, derivative with respect to parameter, pdf, }
}
\description{
This function computes the cdf, pdf, and associated derivatives
}
\examples{
out = margins(0,2.5,"binomial")
}
\author{
Pavel Krupskii and Bruno N. Remillard, January 20, 2022
}
