    \name{multinomial}
     \docType{data}
     \alias{multinomial}
     \title{Simulated data}
     \description{Simulated clustered data from a Clayton copula with parameter 2, and multinomial margins with 3 levels and parameters  1.0,-1 for level 2 and
0.5,  2 for level 3. Clusters and covariates are included.}
     \usage{data(multinomial)}
     \format{Data frame of numerical values}
     \keyword{datasets}

\examples{
data(multinomial)
}
\keyword{datasets}

