    \name{normal}
     \docType{data}
     \alias{normal}
     \title{Simulated data}
     \description{Simulated clustered data from a Clayton copula with parameter 2, rotation = 90,  and normal margins with 1,-1 for the mean, and sd = 4. Clusters and covariates are included.}
     \usage{data(normal)}
     \format{List of simulated values (y, clu, xm)}

     \keyword{datasets}

\examples{
data(normal)
}
\keyword{datasets}

