% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margins.R
\name{normcpdf}
\alias{normcpdf}
\title{normal cdf/pdf and ders}
\usage{
normcpdf(z, th)
}
\arguments{
\item{z}{vector of responses}

\item{th}{th[,1] is mean, th[,2] is standard deviation  > 0;}
}
\value{
\item{out}{Matrix of conditional cdf, derivative with respect to parameter, pdf, }
}
\description{
This function computes the cdf, pdf, and associated derivatives
}
\examples{
out = normcpdf(2,c(-3,4))
}
\author{
Pavel Krupskii and Bruno N. Remillard, January 20, 2022
}
