    \name{out.normal}
     \docType{data}
     \alias{out.normal}
     \title{EstContinuous object}
     \description{Output of EstContinuous for the simulated clustered data normal.}
     \usage{data(out.normal)}
     \format{Data frame of numerical values}


    \keyword{datasets}


\examples{
data(out.normal)
}
\keyword{datasets}

