    \name{out.poisson}
     \docType{data}
     \alias{out.poisson}
     \title{EstDiscrete object}
     \description{Output of EstDiscrete for the simulated clustered data poisson.}
     \usage{data(out.poisson)}
     \format{Data frame of numerical values}


    \keyword{datasets}


\examples{
data(out.poisson)
}
\keyword{datasets}

