% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcond.R
\name{pcondfgm}
\alias{pcondfgm}
\title{Conditional FGM (B10)}
\usage{
pcondfgm(u, v, cpar)
}
\arguments{
\item{u}{probability}

\item{v}{value of the conditioning variable in (0,1)}

\item{cpar}{copula parameter -1<=cpar<=1}
}
\value{
\item{ccdf}{Conditional cdf}
}
\description{
Conditional FGM (B10)
}
\examples{
pcondfgm(0.5,0.6,0.9)
}
