% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcond.R
\name{pcondfrk}
\alias{pcondfrk}
\title{Conditional Frank (B3)}
\usage{
pcondfrk(u, v, cpar)
}
\arguments{
\item{u}{values at which the cdf is evaluated}

\item{v}{value of the conditioning variable in (0,1)}

\item{cpar}{copula parameter}
}
\value{
\item{ccdf}{Conditional cdf}
}
\description{
Conditional Frank (B3)
}
\examples{
pcondfrk(0.5,0.6,2)
}
