% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcond.R
\name{pcondhr}
\alias{pcondhr}
\title{Conditional Huesler-Reiss (B8)}
\usage{
pcondhr(u, v, cpar)
}
\arguments{
\item{u}{values at which the cdf is evaluated}

\item{v}{value of the conditioning variable in (0,1)}

\item{cpar}{copula parameter >0}
}
\value{
\item{ccdf}{Conditional cdf}
}
\description{
Conditional Huesler-Reiss (B8)
}
\examples{
pcondhr(0.5,0.6,2)
}
