% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcond.R
\name{pcondt}
\alias{pcondt}
\title{Conditional Student}
\usage{
pcondt(u, v, cpar, dfC)
}
\arguments{
\item{u}{values at which the cdf is evaluated}

\item{v}{value of the conditioning variable in (0,1)}

\item{cpar}{copula parameter}

\item{dfC}{degrees of freedom}
}
\value{
\item{ccdf}{Conditional cdf}
}
\description{
Conditional Student is Y2|Y1=y1 ~ t(nu+1,location=rho*y1, sigma(y1)), where here sigma^2 = (1-rho^2)(nu+y1^2)/(nu+1)
}
\examples{
pcondt(0.5,0.6,0.6,15)
}
