% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margins.R
\name{poiscpdf}
\alias{poiscpdf}
\title{Poisson cdf/pdf and ders}
\usage{
poiscpdf(z, th)
}
\arguments{
\item{z}{vector of responses}

\item{th}{values of lambda >0}
}
\value{
\item{out}{Matrix of conditional cdf, derivative with respect to parameter, pdf, }
}
\description{
This function computes the cdf, pdf, and associated derivatives
}
\examples{
out = poiscpdf(0,2.5)
}
\author{
Pavel Krupskii and Bruno N. Remillard, January 20, 2022
}
