    \name{poisson}
     \docType{data}
     \alias{poisson}
     \title{Simulated data}
     \description{Simulated clustered data from a Clayton copula with parameter 2, rotation 90, and poisson margins with 1.0,-0.5. Clusters and covariates are included.}
     \usage{data(poisson)}


\format{List of simulated values (y, clu,xm).}
\keyword{datasets}

\examples{
data(poisson)
}
\keyword{datasets}

