% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudosC.R
\name{pseudosC}
\alias{pseudosC}
\title{Estimation cdf, left-continuous cdf, and pseudo-observations}
\usage{
pseudosC(y)
}
\arguments{
\item{y}{univariate data}
}
\value{
\item{Fn}{Emprirical cdf}

\item{Fm}{Left-contniuous cdf}

\item{U}{Pseudo-obsevations}
}
\description{
This function estimates the empirical cdf, its left limit, and pseudo-observations for a univatiate vector y
}
\examples{
y = rpois(100,2)
out=pseudosC(y)

}
\author{
Bruno N. Remillard, January 20, 2022
}
