% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcond.R
\name{qcond}
\alias{qcond}
\title{Inverse conditional cdf}
\usage{
qcond(w, v, family, cpar, rot = 0)
}
\arguments{
\item{w}{probability}

\item{v}{value of the conditioning variable in (0,1)}

\item{family}{"gaussian" , "t" , "clayton" ,  "fgm",  "frank" , "gumbel", "plackett", "galambos", "huesler-reiss"}

\item{cpar}{copula parameter (vector)}

\item{rot}{rotation: 0 (default), 90, 180 (survival), or 270}
}
\value{
\item{U}{Conditional quantile}

\item{U}{Conditional quantile}
}
\description{
This function computes the quantile of conditional cdf  C(U|v) for a copula C
}
\examples{
U = qcond(0.1,0.2,"gaussian",0.87)
}
\author{
Pavel Krupskii and Bruno N. Remillard, January 20, 2022
}
