% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcond.R
\name{qcondfgm}
\alias{qcondfgm}
\title{Inverse  FGM (B10)}
\usage{
qcondfgm(w, v, th)
}
\arguments{
\item{w}{probability}

\item{v}{value of the conditioning variable in (0,1)}

\item{th}{copula parameter -1<=th<=1}
}
\value{
\item{out}{Conditional quantile}
}
\description{
Inverse  FGM (B10)
}
