% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcond.R
\name{qcondgum}
\alias{qcondgum}
\title{Inverse Gumbel}
\usage{
qcondgum(w, v, th)
}
\arguments{
\item{w}{probability}

\item{v}{value of the conditioning variable in (0,1)}

\item{th}{copula parameter}
}
\value{
\item{out}{Conditional quantile}
}
\description{
Inverse Gumbel
}
