% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcond.R
\name{qcondhr}
\alias{qcondhr}
\title{Inverse Huesler-Reiss}
\usage{
qcondhr(w, v, th)
}
\arguments{
\item{w}{probability}

\item{v}{value of the conditioning variable in (0,1)}

\item{th}{copula parameter >0}
}
\value{
\item{out}{Conditional quantile}
}
\description{
Inverse Huesler-Reiss
}
