% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcond.R
\name{qcondjoe}
\alias{qcondjoe}
\title{Inverse Joe}
\usage{
qcondjoe(w, v, th)
}
\arguments{
\item{w}{probability}

\item{v}{value of the conditioning variable in (0,1)}

\item{th}{copula parameter >-1}
}
\value{
\item{out}{Conditional quantile}
}
\description{
Inverse Joe
}
