% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcond.R
\name{qcondnor}
\alias{qcondnor}
\title{Inverse Gaussian}
\usage{
qcondnor(w, v, th)
}
\arguments{
\item{w}{probability}

\item{v}{value of the conditioning variable in (0,1)}

\item{th}{copula parameter (correlation)}
}
\value{
\item{out}{Conditional quantile}
}
\description{
Inverse Gaussian
}
