    \name{sim.poisson}
     \docType{data}
     \alias{sim.poisson}
     \title{Simulated data}
     \description{Simulated clustered data from a Frank copula with parC=c(2,8),  and Poisson margins with parM=c(3.0,-0.1). Clusters and covariates (both uniform) are included.}
     \usage{data(sim.poisson)}


\format{List of simulated values (y, clu,xc,xm) together with true parameters}
\keyword{datasets}

\examples{
data(sim.poisson)
}
\keyword{datasets}

