% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margins.R
\name{tcpdf}
\alias{tcpdf}
\title{Student  cdf/pdf and ders}
\usage{
tcpdf(z, th, df)
}
\arguments{
\item{z}{vector of responses}

\item{th}{th[,1] is mean, th[,2] is standard deviation  > 0}

\item{df}{degrees of freedom}
}
\value{
\item{out}{Matrix of conditional cdf, derivative with respect to parameter, pdf, }
}
\description{
This function computes the cdf, pdf, and associated derivatives
}
\examples{
out = tcpdf(2,c(-3,4),25)
}
\author{
Pavel Krupskii and Bruno N. Remillard, January 20, 2022
}
