% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margins.R
\name{weibcpdf}
\alias{weibcpdf}
\title{Weibul cdf/pdf and ders}
\usage{
weibcpdf(z, th)
}
\arguments{
\item{z}{vector of responses}

\item{th}{th[,1] is rate>0, th[,2] is shape  > 0;}
}
\value{
\item{out}{Matrix of conditional cdf, derivative with respect to parameter, pdf, }
}
\description{
This function computes the cdf, pdf, and associated derivatives
}
\examples{
out = weibcpdf(2,c(2,3))
}
\author{
Pavel Krupskii and Bruno N. Remillard, January 20, 2022
}
